Packager parsing unit tests (UNFINISHED). I have created ConfigParsingTest.java as a generic black box test for parsing the config.xml. ConfigXMLParserTest.java tests the individual methods used for parsing (UNFINISHED) in the ConfigXMLParser class.
As a side note, unit testing the existing packager code has proved to be more difficult than initially anticipated.
The code was not designed for unit testing. The parsing code is contained in several private methods and is tightly bound to neighboring classes.
This made it extremely difficult to unit test properly...

I have completed what I can in the allotted time for this sprint and am submitting what I have so far.